Save A sale feature is Complete!
On saving a Sales the Max required stock for a particular product gets
updated by subtracting the Units a user is selling from the Max required
stock such that the next time the user Makes a sale for the same product
the number of max required units is the updated value.

Like this while saving a sale; two operations we are performing--
inserting into t_sales table and up-dation into t_product table.

All these are happening as a Transaction; either all the 2 operation
happen or none!