I have utilized @ResponseBody to access data in my REST controller method-saveEmployee $.ajax({           type: "POST",           url: "/JsonHtml/emp/saveEmp",           contentType: "application/json",           data: JSON.stringify(formData),           success: function (response) {             alert("Employee saved successfully");           },           error: function (xhr, status, error) {             alert("Error saving employee: " + JSON.stringify(formData));             console.log(JSON.stringify(formData));           },         });
In the View to get this data, I have utilized a simple Html form.
To grab this data from the Form from the User,I have utilized an Ajax
call to send Json data to the Controller endpoint of saveEmployee.

Notice how the keys: contentType and data look like in the AJax.